+++ src/backends/RtAudio/AudioDestinationRtAudio.cpp
@@ -35,7 +35,7 @@ static int NumDefaultOutputChannels()
 const float kLowThreshold = -1.0f;
 const float kHighThreshold = 1.0f;
 
-AudioDestination * AudioDestination::MakePlatformAudioDestination(AudioIOCallback & callback, unsigned numberOfOutputChannels, float sampleRate)
+AudioDestination * AudioDestination::MakePlatformAudioDestination(AudioIOCallback & callback, size_t numberOfOutputChannels, float sampleRate)
 {
     return new AudioDestinationRtAudio(callback, numberOfOutputChannels, sampleRate);
 }
@@ -45,7 +45,7 @@ unsigned long AudioDestination::maxChannelCount()
     return NumDefaultOutputChannels();
 }
 
-AudioDestinationRtAudio::AudioDestinationRtAudio(AudioIOCallback & callback, unsigned numChannels, float sampleRate)
+AudioDestinationRtAudio::AudioDestinationRtAudio(AudioIOCallback & callback, size_t numChannels, float sampleRate)
 : m_callback(callback)
 , m_renderBus(numChannels, AudioNode::ProcessingSizeInFrames, false)
 , m_inputBus(1, AudioNode::ProcessingSizeInFrames, false)
@@ -58,7 +58,6 @@ AudioDestinationRtAudio::AudioDestinationRtAudio(AudioIOCallback & callback, uns
 
 AudioDestinationRtAudio::~AudioDestinationRtAudio()
 {
-    //dac.release(); // XXX
      if (dac.isStreamOpen())
         dac.closeStream();
 }
